refactor(Modal, Overflow, Takeover): WIP
